Move GeoJsonResultObject to geojson package
